requirements.txt
discord.py
requests
flask
pymongo
